// frontend-legacy/src/app/compliance/shared/enums/compliance-params.enum.ts
export enum ComplianceParamsEnum {
  PAGE = 'page',
  SIZE = 'size',
  FROM = 'from',
  TIMESTAMP = '@timestamp',
  TO = 'to',
  TOP = 'top',
  TEMPLATE = 'template',
  VISUALIZATION = 'visualizationsId',
  DATA_ORIGIN = 'dataOrigin',
  DASHBOARD = 'dashboard',
  STANDARD_ID = 'standardId',
  SECTION_ID = 'sectionId',
  CUSTOM = 'codeCompliance',
  STANDARD_NAME = 'standardName'
}